MessageBuilder Refactor
