+++ src/controllers/Interactions.ts
@@ -1,10 +1,11 @@
 import { ButtonInteraction, Message, TextChannel } from "discord.js";
 import { joinQueue, leaveQueue } from "../services/QueueService";
-import MessageBuilder, { ButtonCustomID } from "../utils/MessageBuilder";
+import MessageBuilder from "../utils/MessageHelper/MessageBuilder";
 import { createRandomMatch } from "../services/MatchService";
 import { PlayerInQueue } from "../repositories/QueueRepository/types";
 import QueueRepository from "../repositories/QueueRepository";
 import { setCaptains } from "../services/TeamAssignmentService";
+import ButtonBuilder, { ButtonCustomID } from "../utils/MessageHelper/ButtonBuilder";
 
 export async function postCurrentQueue(queueChannel: TextChannel): Promise<void> {
   const ballchasers = await QueueRepository.getAllBallChasersInQueue();
@@ -17,7 +18,7 @@ export async function handleInteraction(buttonInteraction: ButtonInteraction): P
 
   switch (buttonInteraction.customId) {
     case ButtonCustomID.JoinQueue: {
-      await message.edit(MessageBuilder.disabledQueueButtons(buttonInteraction));
+      await message.edit(ButtonBuilder.disabledQueueButtons(buttonInteraction));
 
       const ballchasers = await joinQueue(buttonInteraction.user.id, buttonInteraction.user.username);
 
@@ -28,20 +29,20 @@ export async function handleInteraction(buttonInteraction: ButtonInteraction): P
           await message.edit(MessageBuilder.queueMessage(ballchasers));
         }
       } else {
-        await message.edit(MessageBuilder.enabledQueueButtons());
+        await message.edit(ButtonBuilder.enabledQueueButtons());
       }
 
       break;
     }
 
     case ButtonCustomID.LeaveQueue: {
-      await message.edit(MessageBuilder.disabledQueueButtons(buttonInteraction));
+      await message.edit(ButtonBuilder.disabledQueueButtons(buttonInteraction));
       const remainingMembers = await leaveQueue(buttonInteraction.user.id);
 
       if (remainingMembers) {
         await message.edit(MessageBuilder.queueMessage(remainingMembers));
       } else {
-        await message.edit(MessageBuilder.enabledQueueButtons());
+        await message.edit(ButtonBuilder.enabledQueueButtons());
       }
       break;
     }
@@ -74,7 +75,7 @@ export async function handleInteraction(buttonInteraction: ButtonInteraction): P
         const ballChasers = await QueueRepository.getAllBallChasersInQueue();
         const players = await setCaptains(ballChasers);
 
-        await message.edit(MessageBuilder.blueChooseMessage(players));
+        await message.edit(MessageBuilder.captainChooseMessage(true, players));
         break;
       }
     }
