+++ src/controllers/MenuInteractions.ts
@@ -3,8 +3,8 @@ import QueueRepository from "../repositories/QueueRepository";
 import { createMatchFromChosenTeams } from "../services/MatchService";
 import { bluePlayerChosen, orangePlayerChosen } from "../services/TeamAssignmentService";
 import { Team } from "../types/common";
-import getEnvVariable from "../utils/getEnvVariable";
-import MessageBuilder, { MenuCustomID } from "../utils/MessageBuilder";
+import { getEnvVariable } from "../utils";
+import MessageBuilder, { MenuCustomID } from "../utils/MessageHelper/MessageBuilder";
 
 export async function handleMenuInteraction(menuInteraction: SelectMenuInteraction): Promise<void> {
   const message = menuInteraction.message;
@@ -19,7 +19,7 @@ export async function handleMenuInteraction(menuInteraction: SelectMenuInteracti
       if (isCaptain || isDev) {
         const playersLeft = await bluePlayerChosen(menuInteraction.values[0]);
 
-        await message.edit(MessageBuilder.orangeChooseMessage(playersLeft));
+        await message.edit(MessageBuilder.captainChooseMessage(false, playersLeft));
         break;
       }
